Create Aeotec Garage Door Controller Gen 5
Robert Vandervoort's device type for the controller, made for smartthings, adapted for hubitat. 

Configure Audio on lines 422, 429, 433, 437, 440
According to this logic
  
(blinking)(alarm sound)(volume)(enable/disable)

The first value is the amount of blinking, from 1 to 10. 
The alarm sound is one you have uploaded to the device, with a corresponding track number. 
The volume of the sound, 1 to 8. 
Whether to play a sound or not. 0 to disable, 1 to enable. 

https://aeotec.freshdesk.com/support/solutions/articles/6000142866-configure-garage-door-controller-alarm-sounds-